+++ test/style_trans.js
@@ -30,14 +30,14 @@ suite('style_trans', function() {
 "#tab[zoom=1] { marker-width:10; marker-height:20; }\n#tab[zoom=2] { marker-height:'6'; marker-width: '7'; }"
     , '2.0.2', '2.1.0'
     );
-    var e = "#tab[zoom=1] { marker-width:20; marker-height:40; [mapnik-geometry-type=1] { marker-placement:point; marker-type:ellipse; } [mapnik-geometry-type>1] { marker-placement:line; marker-type:arrow; marker-clip:false; } }\n#tab[zoom=2] { marker-height:'12'; marker-width:'14'; [mapnik-geometry-type=1] { marker-placement:point; marker-type:ellipse; } [mapnik-geometry-type>1] { marker-placement:line; marker-type:arrow; marker-clip:false; } }"
+    var e = "#tab[zoom=1] { marker-width:20; marker-height:40; [mapnik-geometry-type=1] { marker-placement:point; marker-type:ellipse; } [mapnik-geometry-type>1] { marker-placement:line; marker-type:arrow; marker-transform:scale(.5, .5); marker-clip:false; } }\n#tab[zoom=2] { marker-height:'12'; marker-width:'14'; [mapnik-geometry-type=1] { marker-placement:point; marker-type:ellipse; } [mapnik-geometry-type>1] { marker-placement:line; marker-type:arrow; marker-transform:scale(.5, .5); marker-clip:false; } }"
     assert.equal(s,e);
 
     var s = t.transform(
 "#t { marker-width :\n10; \nmarker-height\t:   20; }"
     , '2.0.2', '2.1.0'
     );
-    var e = "#t { marker-width:20; \nmarker-height:40; [mapnik-geometry-type=1] { marker-placement:point; marker-type:ellipse; } [mapnik-geometry-type>1] { marker-placement:line; marker-type:arrow; marker-clip:false; } }"
+    var e = "#t { marker-width:20; \nmarker-height:40; [mapnik-geometry-type=1] { marker-placement:point; marker-type:ellipse; } [mapnik-geometry-type>1] { marker-placement:line; marker-type:arrow; marker-transform:scale(.5, .5); marker-clip:false; } }"
     assert.equal(s,e);
 
   });
@@ -50,10 +50,10 @@ suite('style_trans', function() {
     );
     var e = "#t [a<1] { marker-width:2; "
           + "[mapnik-geometry-type=1] { marker-placement:point; marker-type:ellipse; } "
-          + "[mapnik-geometry-type>1] { marker-placement:line; marker-type:arrow; marker-clip:false; } }"
+          + "[mapnik-geometry-type>1] { marker-placement:line; marker-type:arrow; marker-transform:scale(.5, .5); marker-clip:false; } }"
           + " # [a>1] { marker-width:4; "
           + "[mapnik-geometry-type=1] { marker-placement:point; marker-type:ellipse; } "
-          + "[mapnik-geometry-type>1] { marker-placement:line; marker-type:arrow; marker-clip:false; } }";
+          + "[mapnik-geometry-type>1] { marker-placement:line; marker-type:arrow; marker-transform:scale(.5, .5); marker-clip:false; } }";
 //console.log("O:"+s);
 //console.log("E:"+e);
     assert.equal(s, e);
@@ -67,14 +67,14 @@ suite('style_trans', function() {
     );
     var e = "#t[a=1] { marker-width:2; "
           + "[mapnik-geometry-type=1] { marker-placement:point; marker-type:ellipse; } "
-          + "[mapnik-geometry-type>1] { marker-placement:line; marker-type:arrow; marker-clip:false; } } "
+          + "[mapnik-geometry-type>1] { marker-placement:line; marker-type:arrow; marker-transform:scale(.5, .5); marker-clip:false; } } "
           + "#t[a=2] { line-color:red; line-clip:false; } "
           + "#t[a=3] { marker-placement:line; "
           // NOTE: we do override marker-placement for points because "line" doesn't work in 2.1.0
           //       and it worked exactly as "point" in 2.0.0
           + "[mapnik-geometry-type=1] { marker-placement:point; marker-type:ellipse; } "
           // NOTE: we do NOT override marker-placement for lines or polys
-          + "[mapnik-geometry-type>1] { marker-type:arrow; marker-clip:false; } }"
+          + "[mapnik-geometry-type>1] { marker-type:arrow; marker-transform:scale(.5, .5); marker-clip:false; } }"
         ;
 //console.log("O:"+s);
 //console.log("E:"+e);
@@ -87,7 +87,7 @@ suite('style_trans', function() {
 "#tab[zoom=1] { marker-width:10; marker-height:20; }\n#tab[zoom=2] { marker-height:'6'; marker-width: \"7\"; }"
     , '2.0.0', '2.1.0'
     );
-    var e = "#tab[zoom=1] { marker-width:20; marker-height:40; [mapnik-geometry-type=1] { marker-placement:point; marker-type:ellipse; } [mapnik-geometry-type>1] { marker-placement:line; marker-type:arrow; marker-clip:false; } }\n#tab[zoom=2] { marker-height:'12'; marker-width:\"14\"; [mapnik-geometry-type=1] { marker-placement:point; marker-type:ellipse; } [mapnik-geometry-type>1] { marker-placement:line; marker-type:arrow; marker-clip:false; } }"
+    var e = "#tab[zoom=1] { marker-width:20; marker-height:40; [mapnik-geometry-type=1] { marker-placement:point; marker-type:ellipse; } [mapnik-geometry-type>1] { marker-placement:line; marker-type:arrow; marker-transform:scale(.5, .5); marker-clip:false; } }\n#tab[zoom=2] { marker-height:'12'; marker-width:\"14\"; [mapnik-geometry-type=1] { marker-placement:point; marker-type:ellipse; } [mapnik-geometry-type>1] { marker-placement:line; marker-type:arrow; marker-transform:scale(.5, .5); marker-clip:false; } }"
 //console.log("O:"+s);
 //console.log("E:"+e);
     assert.equal(s, e, "Obt:"+s+"\nExp:"+s);
@@ -96,21 +96,21 @@ suite('style_trans', function() {
 "#t { marker-width :\n10; \nmarker-height\t:   20; }"
     , '2.0.0', '2.1.0'
     );
-    var e = "#t { marker-width:20; \nmarker-height:40; [mapnik-geometry-type=1] { marker-placement:point; marker-type:ellipse; } [mapnik-geometry-type>1] { marker-placement:line; marker-type:arrow; marker-clip:false; } }"
+    var e = "#t { marker-width:20; \nmarker-height:40; [mapnik-geometry-type=1] { marker-placement:point; marker-type:ellipse; } [mapnik-geometry-type>1] { marker-placement:line; marker-type:arrow; marker-transform:scale(.5, .5); marker-clip:false; } }"
     assert.equal(s, e);
 
     var s = t.transform(
 "#tab { marker-width:2 }"
     , '2.0.0', '2.1.0'
     );
-    var e = "#tab { marker-width:4; [mapnik-geometry-type=1] { marker-placement:point; marker-type:ellipse; } [mapnik-geometry-type>1] { marker-placement:line; marker-type:arrow; marker-clip:false; } }"
+    var e = "#tab { marker-width:4; [mapnik-geometry-type=1] { marker-placement:point; marker-type:ellipse; } [mapnik-geometry-type>1] { marker-placement:line; marker-type:arrow; marker-transform:scale(.5, .5); marker-clip:false; } }"
     assert.equal(s, e);
 
     var s = t.transform(
 "#tab{ marker-width:2 }"
     , '2.0.0', '2.1.0'
     );
-    var e = "#tab{ marker-width:4; [mapnik-geometry-type=1] { marker-placement:point; marker-type:ellipse; } [mapnik-geometry-type>1] { marker-placement:line; marker-type:arrow; marker-clip:false; } }";
+    var e = "#tab{ marker-width:4; [mapnik-geometry-type=1] { marker-placement:point; marker-type:ellipse; } [mapnik-geometry-type>1] { marker-placement:line; marker-type:arrow; marker-transform:scale(.5, .5); marker-clip:false; } }";
     assert.equal(s, e);
 
   });
@@ -129,7 +129,7 @@ suite('style_trans', function() {
 "#t{ marker-line-color:red; }"
     , '2.0.0', '2.1.0'
     );
-    var e = "#t{ marker-line-color:red; [mapnik-geometry-type=1] { marker-placement:point; marker-type:ellipse; } [mapnik-geometry-type>1] { marker-placement:line; marker-type:arrow; marker-clip:false; } }"
+    var e = "#t{ marker-line-color:red; [mapnik-geometry-type=1] { marker-placement:point; marker-type:ellipse; } [mapnik-geometry-type>1] { marker-placement:line; marker-type:arrow; marker-transform:scale(.5, .5); marker-clip:false; } }"
     assert.equal(s, e);
   });
 
@@ -184,7 +184,7 @@ suite('style_trans', function() {
 "#t{ marker-fill-color:red; }"
     , '2.0.0', '2.1.1'
     );
-    var e = "#t{ marker-fill-color:red; [mapnik-geometry-type=1] { marker-placement:point; marker-type:ellipse; } [mapnik-geometry-type>1] { marker-placement:line; marker-type:arrow; marker-clip:false; } marker-multi-policy:whole; }"
+    var e = "#t{ marker-fill-color:red; [mapnik-geometry-type=1] { marker-placement:point; marker-type:ellipse; } [mapnik-geometry-type>1] { marker-placement:line; marker-type:arrow; marker-transform:scale(.5, .5); marker-clip:false; } marker-multi-policy:whole; }"
     assert.equal(s, e);
   });